refactor: move Smart Detector alert rules to separate resources (#58)
No need for a `for_each` when we create a static number of Smart Detector alert rule resources.

Currently, the Terraform Registry is not able to count the number of resources created when using `for_each`. If my change is merged, the Terraform Registry should explicitly list 6 `azurerm_monitor_smart_detector_alert_rule` resources to be created.

Release-As: 5.3.3